pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "datetime>=5.5",
    "python-binance>=1.0.29",
    "requests>=2.32.5",
]
